move backgroundsync related stuff from SettingsActivity to ZerdaReaderApp
