+++ app/src/main/java/com/greenfox/fuchsit/zerdareader/activity/SettingsActivity.java
@@ -1,20 +1,15 @@
 package com.greenfox.fuchsit.zerdareader.activity;
 
-import android.app.AlarmManager;
-import android.app.PendingIntent;
-import android.content.Context;
-import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.preference.PreferenceActivity;
 import android.preference.PreferenceManager;
-import android.util.Log;
 
 import com.greenfox.fuchsit.zerdareader.R;
 import com.greenfox.fuchsit.zerdareader.ZerdaReaderApp;
-import com.greenfox.fuchsit.zerdareader.backgroundSync.BackgroundSyncReceiver;
+import com.greenfox.fuchsit.zerdareader.event.BackSyncSettingEvent;
 
-import java.util.Set;
+import org.greenrobot.eventbus.EventBus;
 
 public class SettingsActivity extends PreferenceActivity implements SharedPreferences.OnSharedPreferenceChangeListener {
     
@@ -34,42 +29,12 @@ public class SettingsActivity extends PreferenceActivity implements SharedPrefer
     public void onSharedPreferenceChanged(SharedPreferences loginData, String key) {
         setPreferenceScreen(null);
         addPreferencesFromResource(R.xml.preferences);
-        enableBackgroundSync(preferences.getBoolean("checkbox_sync", true));
+        EventBus.getDefault().post(new BackSyncSettingEvent(preferences.getBoolean("checkbox_sync", true)));
     }
 
-    public void enableBackgroundSync(boolean checkboxChecked) {
-        if(checkboxChecked) {
-            Log.e("SettingsActivity", "background sync enabled");
-            scheduleAlarm();
-        } else {
-            Log.e("SettingsActivity", "background sync disabled");
-            cancelAlarm();
-        }
-    }
-
-    public void scheduleAlarm() {
-        final PendingIntent pIntent = setupPendingIntent();
-        AlarmManager alarm = (AlarmManager) this.getSystemService(Context.ALARM_SERVICE);
-        alarm.setInexactRepeating(AlarmManager.RTC_WAKEUP, System.currentTimeMillis(), defineInterval(), pIntent);
-    }
-
-    private long defineInterval() {
-        if(ZerdaReaderApp.inForeground) {
-            Log.e("SettingsAct", "in Foreground");
-            return 120000L;
-        } else {
-            Log.e("SettingsAct", "not in Foreground");
-            return 600000L;
-        }
-    }
-
-    public void cancelAlarm() {
-        AlarmManager alarm = (AlarmManager) this.getSystemService(Context.ALARM_SERVICE);
-        alarm.cancel(setupPendingIntent());
-    }
-
-    public PendingIntent setupPendingIntent() {
-        Intent intent = new Intent(this, BackgroundSyncReceiver.class);
-        return PendingIntent.getBroadcast(this, 0, intent, PendingIntent.FLAG_UPDATE_CURRENT);
+    @Override
+    protected void onResume() {
+        super.onResume();
+        ZerdaReaderApp.visible = true;
     }
 }
